agent_workspace/polymarket_arbitrage_bot/requirements.txt
# 核心依赖
python>=3.8

# HTTP请求
requests>=2.28.0
aiohttp>=3.8.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0

# 配置管理
pyyaml>=6.0
python-dotenv>=0.21.0

# 日志记录
loguru>=0.7.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0

# 异步处理
asyncio>=3.4.3

# WebSocket支持
websockets>=11.0.0

# 测试框架
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# 代码质量
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# 类型提示
typing-extensions>=4.5.0

# 日期时间处理
python-dateutil>=2.8.2

# 进度条
tqdm>=4.65.0

# 数据缓存
redis>=4.5.0

# 数据库支持
sqlalchemy>=2.0.0
alembic>=1.11.0

# API文档生成
fastapi>=0.95.0
uvicorn>=0.21.0